// webtool/src/types/ArtistData.ts
interface ArtistData {
    creatorbirthanddeathdescription: string;
    creatorbirthdate: number;
    creatordeathdate: number;
    creatorfirstname: string;
    creatorfulldescription: string;
    creatorid: string;
    creatorlastname: string;
    creatornationality: string;
    artworkrecordids: number[];
}

export default ArtistData;